Refactor CLI commands into dedicated module and improve command handling
