+++ openhands/core/cli_settings.py
@@ -6,8 +6,8 @@ from prompt_toolkit.widgets import Frame, TextArea
 from pydantic import SecretStr
 
 from openhands.controller.agent import Agent
+from openhands.core.cli_input import UserCancelledError, cli_confirm, kb_cancel
 from openhands.core.cli_output import COLOR_GREY
-from openhands.core.cli_prompts import UserCancelledError, cli_confirm, kb_cancel
 from openhands.core.cli_utils import (
     VERIFIED_ANTHROPIC_MODELS,
     VERIFIED_OPENAI_MODELS,
